requirements.txt
pylint
requests
python-dotenv
websocket-client==0.56
-e git+git://github.com/carlosmds/iqoptionapi#egg=iqoptionapi